Create Graph structure, improve path lookups
The Graph structure now enables printing only the templates that have no
parents, which are expected to be the "root" templates rendered by a
controller.

Additionally, paths are now normalized to absolute paths so that the
Graph is able to show the entire tree of partials rendered by a root
template.