Use view_paths for template load/lookup
Previously only the app's template were loaded, which isn't sufficient
for apps that mount engines with views.

This commit now boots the app so that it can load templates from all
view_paths. This will likely be an overcorrection and begin to show
loose templates from gems, but that can be resolved once the output
focuses on just the loose templates.