Print _all_ loose ERBs by default
Previously it would print ERBs without any parents, but this output
isn't ideal because it isn't clear which dependency of a loose partial
is also loose (it could be rendered by some other template).

Now ERBs are printed if they are unused, or all of their parents are
unused. This is more actionable because every ERB printed should be a
candiate for deletion.