Add scanner for renders in Helper modules
This helps to remove some false positives where a partial was thought to
be loose because it wasn't rendered by a template, but in fact it was
rendered by a Helper method.

This could result in false negatives if a Helper method is actually
unused, but being able to find regular unused Ruby methods is out of
scope for this project.