Refactor some names to improve intention
`used_erbs` is not terribly descriptive since the `#select` uses other
factors to determine whether an erb is "used". `ruby_rendered_erbs`
better indicates that this is the list of erbs explicitly rendered in a
helper (and controllers/etc. in the future)

`LooseVisitor` is also a bit confusing because its actually walking
dependency trees to mark erbs as `not_loose!`, so `NotLooseVisitor`
is a better name.