Pass a real LookupContext to the Registry
On the path towards using more built-in Rails classes for more accurate
resolution, this commit tweaks the `Registry` to accept a
`LookupContext` instead of an array of view paths.

The `LookupContext` will contain the exact same `view_paths` as before,
but it also now gives us access to Rails' erb lookup strategy. It also
makes the tests bit cleaner, since we can use a real class instead an
undocumented Rails class (`ActionView::Pathset`).