Move all `loose?` logic into the `LooseVisitor`
`ActionView::Digestor` has its own `Node` implementation, so any fields
that can be moved off of `LooseErbs`'s `Node` will make a clean
migration easier and avoid having to monkeypatch.

Since the `loose?` logic was moved to the visitor, it additionally
required a small refactor to the `LooseFilter`. It was also changed to
only run when it would be used (when not `--all`), which exposed that a
few places were relying on `nodes` being an `Array` and not a generic
`Enumerable`.