Replace `Registry#dependencies_for` with `#tree`
This replaces the custom Node/Graph/dependency code with
`ActionView::Digestor`. `Digestor`'s implementation is basically a
superset of the previous functionality: it does all the same things as
before but handles some things better.

One small change is that now `Missing` nodes get their own class and
they are added to the `Registry` when found. This is a bit better than
before, as the previous implementation would warn each time a "Missing"
Node would have been encountered. Now the warnings can be collected so
that there is only one per `Missing` node name, and they can be
optionally printed at any time.