Use the LookupContext in Registry#lookup
Previously, the lookup method was written at a time where I was writing
most of the gem from scratch and not using the Digestor for tree
generation (and I even indicated as such with a comment that the method
is a hack).

This commit refactors the method to actually use Action View's
LookupContext since all the pieces are already present. This should
generally provide more accurate template lookups than the code before
since its using even more of the logic used by Rails.

While doing that refactor, I also refactored the Cli a bit (lots more
helper methods to make comments unnecessary and/or replacing local
variables).